update login & register

+++ src/main/resources/static/js/loginOrRegister.js
@@ -122,7 +122,7 @@ function pwdValid(pwd, tip) {
 }
 
 //pwd & rePwd check
-function checkPwdMatch(pwd, rePwd, rePwdTipTxt) {
+function isPwdMatch(pwd, rePwd, rePwdTipTxt) {
     if (pwd.value !== rePwd.value) {
         rePwdTipTxt.textContent = "新密碼與再次輸入不同";
         return false;
@@ -175,8 +175,16 @@ btnLogin.addEventListener("click", function () {
             }),
         }).then(res => res.json()).then(data => {
             if (data.successful) {
-                sessionStorage.setItem("memberData", JSON.stringify(data));
-                location.href = "../homePage.html";
+                Swal.fire({
+                    icon: "success",
+                    title: "登入成功",
+                    showConfirmButton: false,
+                    timer: 1500
+                }).then(() => {
+                    sessionStorage.setItem("memberData", JSON.stringify(data));
+
+                    location.replace("../homePage.html");
+                })
             } else {
                 Swal.fire(data.message, "", "error");
             }
@@ -190,7 +198,7 @@ btnRegister.addEventListener("click", function () {
         Swal.fire("請輸入手機 & 密碼 & 姓名", "", "error");
     } else if (!(phoneValid(memberPhoneInput) && pwdValid(pwdInput, pwdTipTxt) && pwdValid(rePwdInput, rePwdTipTxt))) {
         Swal.fire("手機與密碼有誤，請重新確認", "", "error");
-    } else if (!checkPwdMatch(pwdInput, rePwdInput, rePwdTipTxt)) {
+    } else if (!isPwdMatch(pwdInput, rePwdInput, rePwdTipTxt)) {
         Swal.fire("新密碼與再次輸入不同", "", "error");
     } else {
         fetch(`member/register`, {
@@ -203,13 +211,11 @@ btnRegister.addEventListener("click", function () {
             }),
         }).then(res => res.json()).then(data => {
             if (data.successful) {
-                console.log(1);
+                // todo - jwt
                 sessionStorage.setItem("memberData", JSON.stringify(data.data));
-                location.href = "../memberCertificate.html";
+                location.replace("../memberCertificate.html");
             } else {
-                console.log(data.message);
-
-                Swal.fire(data.message == null ? "" : data.message, "", "error");
+                Swal.fire(data.message == null ? "註冊失敗" : data.message, "", "error");
             }
         })
     }
